+++ CHANGELOG.md
@@ -22,6 +22,10 @@ That said, these are more guidelines rather than hard rules, though the project 
 
 ## 0.14.7/0.15.0 - Unreleased
 
+### Features
+
+- [#2174](https://github.com/ClementTsang/bottom/pull/2174): Support setting default sort order in config.
+
 ### Other
 
 - [#2167](https://github.com/ClementTsang/bottom/pull/2167): Remove filedescriptor hack used to suppress false-positive stderr messages for older versions of FreeBSD.